+++ tests/unit_tests_old/tests_s_nb/test_notebooks.py
@@ -7,7 +7,7 @@ import sys
 import wandb
 
 from wandb.errors import UsageError
-from tests.unit_tests import utils
+from tests.unit_tests_old import utils
 
 
 def test_login_timeout(notebook, monkeypatch):
